refactor: reset project with proper apiversioning
This resets the whole project to fix a mistake in the earlier setup
when specifying the group name (no need to include domain in the
operator-sdk create api group command).

I also deleted all the manifest related information. I don't need
those for now, but it may become useful in the future when
distributing this.